+++ test/api/storage/adapter.registry.spec.ts
@@ -7,9 +7,9 @@ import {StorageRegistry} from "../../../src/api/storage/adapter.registry";
 function mockStorageAdapter(uri: URI): StorageAdapter {
     /* tslint:disable: no-empty */
     return {
-        loadPage: (_): Promise<PageOverlay> => Promise.resolve(new PageOverlay(1, [], [])),
+        loadPage: (_, __): Promise<PageOverlay> => Promise.resolve(new PageOverlay(1, [], [])),
         register: (): URI => uri,
-        start: (_): void => {}
+        start: (_, __): void => {}
     };
     /* tslint:enable */
 }